Replacing 1-n relational child documents does not always delete:
- When relacing a 1-n relation the existing child documents will now
  follow the proper cascade strategy that was set on the relation:
  :delete, :destroy, :nullify.

- If a dependent strategy was not defined, we assume nullify.

- Fixes #1173.

+++ lib/mongoid/relations/referenced/many.rb
@@ -245,10 +245,14 @@ module Mongoid #:nodoc:
         #
         # @since 2.0.0.beta.1
         def purge
-          criteria.delete_all
-          target.clear do |doc|
-            unbind_one(doc)
-            doc.destroyed = true
+          unless metadata.destructive?
+            nullify
+          else
+            criteria.delete_all
+            target.clear do |doc|
+              unbind_one(doc)
+              doc.destroyed = true
+            end
           end
         end
         alias :clear :purge
